WIP: Update BOSH Agent handles disabling Windows jobs
We discovered that the Windows BOSH Agent does not properly disable BOSH
jobs. It leaves behind the enabled symlink directory
(/var/vcap/jobs/job_name) even though the actual job is uninstalled from
the system. This leaves behind a symlink that points to nothing. This
was due to the Agent comparing Windows style directory names (from
ReadLink/ReadAndFollowLink) to Unix style directory names (from internal
stored data)

This commit contains exploratory changes that comes after updating the
BOSH Utils ReadAndFollowLink and ReadLink.

[#162161084]

NOTE: This commit is not intended to be final code. This is just used to
store save the current investigation so it is not lost.

Signed-off-by: Malini Valliath <mvalliath@pivotal.io>
Co-authored-by: Malini Valliath <mvalliath@pivotal.io>